修正x64版本下webconsole的spaceview功能异常的问题 https://github.com/kbengine/kbengine/issues/469

+++ kbe/src/server/cellapp/space_viewer.cpp
@@ -206,7 +206,7 @@ void SpaceViewer::initClient()
 
 	// 先下发脚本ID对应脚本模块的名称，便于降低后面实体同步量，实体只同步id过去
 	const EntityDef::SCRIPT_MODULES& scriptModules = EntityDef::getScriptModules();
-	s << scriptModules.size();
+	s << (uint32)scriptModules.size();
 
 	EntityDef::SCRIPT_MODULES::const_iterator moduleIter = scriptModules.begin();
 	for (; moduleIter != scriptModules.end(); ++moduleIter)
